refactor: useNumberKeyboard로 기능 관련 역할 분리
